# src/client/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DUCC Website API",
    "description": "API documentation for the Durham University Canoe Club website.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie used for authentication."
      }
    }
  },
  "paths": {
...
    "/api/admin/users": {
        "get": {
            "summary": "List users (Admin)",
            "description": "Retrieves a list of all users with administrative details.",
            "tags": ["Admin"],
            "security": [{ "cookieAuth": [] }],
            "responses": { "200": { "description": "List of users" } }
        }
    },
    "/api/admin/user/{id}": {
        "get": {
            "summary": "Get user details (Admin)",
            "description": "Retrieves detailed information about a specific user for administration.",
            "tags": ["Admin"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the user.", "schema": { "type": "integer" } }],
            "responses": { "200": { "description": "User details" } }
        }
    },
    "/api/admin/user/{id}/elements": {
        "post": {
            "summary": "Update user elements (Admin)",
            "description": "Updates specific profile elements of a user as an administrator.",
            "tags": ["Admin"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the user.", "schema": { "type": "integer" } }],
            "requestBody": { "content": { "application/json": { "schema": { "type": "object", "description": "Object containing fields to update." } } } },
            "responses": { "200": { "description": "User updated" } }
        }
    },
    "/api/admin/user/{id}/role": {
        "post": {
            "summary": "Assign role to user",
            "description": "Assigns a specific role (e.g., Exec, Coach) to a user.",
            "tags": ["Admin"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the user.", "schema": { "type": "integer" } }],
            "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "roleId": { "type": "integer", "description": "The ID of the role to assign." } } } } } },
            "responses": { "200": { "description": "Role assigned" } }
        }
    },
    "/api/admin/user/{id}/role/{roleId}": {
        "delete": {
            "summary": "Remove role from user",
            "description": "Removes a specific role from a user.",
            "tags": ["Admin"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [
                { "name": "id", "in": "path", "required": true, "description": "The ID of the user.", "schema": { "type": "integer" } },
                { "name": "roleId", "in": "path", "required": true, "description": "The ID of the role to remove.", "schema": { "type": "integer" } }
            ],
            "responses": { "200": { "description": "Role removed" } }
        }
    },
    "/api/admin/user/{id}/permission": {
        "post": {
            "summary": "Add permission to user",
            "description": "Grants a specific permission directly to a user.",
            "tags": ["Admin"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the user.", "schema": { "type": "integer" } }],
            "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "permissionId": { "type": "integer", "description": "The ID of the permission to add." } } } } } },
            "responses": { "200": { "description": "Permission added" } }
        }
    },
    "/api/admin/user/{id}/permission/{permId}": {
        "delete": {
            "summary": "Remove permission from user",
            "description": "Revokes a specific permission from a user.",
            "tags": ["Admin"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [
                { "name": "id", "in": "path", "required": true, "description": "The ID of the user.", "schema": { "type": "integer" } },
                { "name": "permId", "in": "path", "required": true, "description": "The ID of the permission to remove.", "schema": { "type": "integer" } }
            ],
            "responses": { "200": { "description": "Permission removed" } }
        }
    },
    "/api/admin/user/{id}/managed_tag": {
        "post": {
            "summary": "Add managed tag scope",
            "description": "Grants a user the ability to manage a specific tag.",
            "tags": ["Admin"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the user.", "schema": { "type": "integer" } }],
            "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "tagId": { "type": "integer", "description": "The ID of the tag to manage." } } } } } },
            "responses": { "200": { "description": "Scope added" } }
        }
    },
    "/api/admin/user/{id}/managed_tag/{tagId}": {
        "delete": {
            "summary": "Remove managed tag scope",
            "description": "Revokes a user's ability to manage a specific tag.",
            "tags": ["Admin"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [
                { "name": "id", "in": "path", "required": true, "description": "The ID of the user.", "schema": { "type": "integer" } },
                { "name": "tagId", "in": "path", "required": true, "description": "The ID of the tag to remove.", "schema": { "type": "integer" } }
            ],
            "responses": { "200": { "description": "Scope removed" } }
        }
    },
    "/api/admin/roles/permissions": {
        "get": {
            "summary": "List all permissions",
            "description": "Retrieves a list of all available permissions in the system.",
            "tags": ["Admin"],
            "security": [{ "cookieAuth": [] }],
            "responses": { "200": { "description": "List of permissions" } }
        }
    },
    "/api/admin/roles": {
        "get": {
            "summary": "List all roles",
            "description": "Retrieves a list of all defined roles in the system.",
            "tags": ["Admin"],
            "security": [{ "cookieAuth": [] }],
            "responses": { "200": { "description": "List of roles" } }
        },
        "post": {
            "summary": "Create role",
            "description": "Creates a new role definition with assigned permissions.",
            "tags": ["Admin"],
            "security": [{ "cookieAuth": [] }],
            "requestBody": { "content": { "application/json": { "schema": { "type": "object", "description": "Role definition object." } } } },
            "responses": { "200": { "description": "Role created" } }
        }
    },
    "/api/admin/roles/{id}": {
        "get": {
            "summary": "Get specific role details",
            "description": "Retrieves the details of a specific role, including its permissions.",
            "tags": ["Admin"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the role.", "schema": { "type": "integer" } }],
            "responses": { "200": { "description": "Role details" } }
        },
        "put": {
            "summary": "Update role",
            "description": "Updates an existing role's name or permissions.",
            "tags": ["Admin"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the role.", "schema": { "type": "integer" } }],
            "requestBody": { "content": { "application/json": { "schema": { "type": "object", "description": "Updated role fields." } } } },
            "responses": { "200": { "description": "Role updated" } }
        },
        "delete": {
            "summary": "Delete role",
            "description": "Deletes a role definition from the system.",
            "tags": ["Admin"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the role to delete.", "schema": { "type": "integer" } }],
            "responses": { "200": { "description": "Role deleted" } }
        }
    },
...
  }
}